pyproject.toml
[project]
name = "mockpath"
version = "0.1.1"
description = "Lightweight HTTP mock server — directory structure as URL paths"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["pyyaml", "click>=8.0"]
license = "MIT"
keywords = ["mock", "http", "server", "testing", "api"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Testing",
    "Topic :: Internet :: WWW/HTTP",
]

[project.scripts]
mockpath = "mockpath.cli:main"

[project.urls]
Homepage = "https://github.com/xuopoj/mockpath"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mockpath"]
